Restructured source files to make sense
Have the moved files show up in /pkg directory
