+++ .gitignore
@@ -1,7 +1,6 @@
 /eclipse-bin/
 /bin
 /build
-/pkg
 .classpath
 .project
 .settings
