+++ example/s3_download/main.go
@@ -5,13 +5,13 @@ package main
 
 import (
 	"fmt"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/s3"
 	"io"
 	"os"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	awsS3 "github.com/aws/aws-sdk-go/service/s3"
-	"go.amzn.com/robomaker/bundle_support/s3"
 )
 
 const (
