+++ pkg/3p/archiver/NOTICE
