+++ pkg/archive/bundle_archive.go
@@ -7,9 +7,9 @@ import (
 	"archive/tar"
 	"errors"
 	"fmt"
-	"go.amzn.com/robomaker/bundle_support/bundle"
-	"go.amzn.com/robomaker/bundle_support/extractors"
-	"go.amzn.com/robomaker/bundle_support/store"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/bundle"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/extractors"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/store"
 	"io"
 )
