+++ pkg/archive/bundle_processor.go
@@ -4,8 +4,8 @@
 package archive
 
 import (
-	"go.amzn.com/robomaker/bundle_support/bundle"
-	"go.amzn.com/robomaker/bundle_support/store"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/bundle"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/store"
 	"io"
 )
