+++ pkg/archive/bundle_processor_test.go
