+++ pkg/archive/bundle_processor_v2.go
@@ -8,9 +8,9 @@ import (
 	"compress/gzip"
 	"encoding/json"
 	"fmt"
-	"go.amzn.com/robomaker/bundle_support/bundle"
-	"go.amzn.com/robomaker/bundle_support/extractors"
-	"go.amzn.com/robomaker/bundle_support/store"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/bundle"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/extractors"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/store"
 	"io"
 	"io/ioutil"
 )
