+++ pkg/bundle/bundle_error.go
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package bundle_support
+package bundle
 
 import "fmt"
