+++ pkg/extractors/extractor.go
@@ -9,7 +9,7 @@ package extractors
 import (
 	"archive/tar"
 	"compress/gzip"
-	"go.amzn.com/robomaker/bundle_support/file_system"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/file_system"
 	"io"
 )
