+++ pkg/extractors/targz_extractor.go
@@ -4,8 +4,8 @@ package extractors
 // SPDX-License-Identifier: Apache-2.0
 
 import (
-	"go.amzn.com/robomaker/bundle_support/3p/archiver"
-	"go.amzn.com/robomaker/bundle_support/file_system"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/3p/archiver"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/file_system"
 	"io"
 )
