+++ pkg/extractors/targz_extractor_test.go
@@ -5,9 +5,9 @@ package extractors
 
 import (
 	"errors"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/file_system"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"go.amzn.com/robomaker/bundle_support/file_system"
 	"testing"
 )
