+++ pkg/store/bundle_store_test.go
@@ -5,10 +5,10 @@ package store
 
 import (
 	"errors"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/extractors"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/file_system"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"go.amzn.com/robomaker/bundle_support/extractors"
-	"go.amzn.com/robomaker/bundle_support/file_system"
 	"os"
 	"path/filepath"
 	"testing"
