+++ pkg/stream/path_to_stream.go
@@ -8,12 +8,12 @@ package stream
 import (
 	"crypto/md5"
 	"fmt"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/file_system"
+	"github.com/aws-robotics/aws-robomaker-bundle-support-library/pkg/s3"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	awsS3 "github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
-	"go.amzn.com/robomaker/bundle_support/file_system"
-	"go.amzn.com/robomaker/bundle_support/s3"
 	"io"
 	"os"
 	"regexp"
